Client can send files (up to a certain size ?)

+++ core/src/main/java/fr/uge/teillardnajjar/chatfusion/core/model/parts/IdentifiedFileChunk.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 public record IdentifiedFileChunk(
     Identifier identifier,
     String filename,
-    int fileSize,
+    long fileSize,
     int fileId,
     ByteBuffer chunk
 ) {
